realizado subida y obtener ficheros

+++ workspaces1/jacalix2/src/main/java/com/jacalix/repo/DocumentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jacalix.model.entity.Document;
 
-public interface DocumentRepository extends CrudRepository<Document, Long> {
+public interface DocumentRepository extends CrudRepository<Document, Integer> {
 
 }
